style: checkbox in balance sheet Asset table.
