+++ src/components/ui/indeterminate-checkbox.tsx
@@ -1,0 +1,30 @@
+import React, { HTMLProps } from 'react'
+
+function IndeterminateCheckbox({
+  indeterminate,
+  className = '',
+  ...rest
+}: { indeterminate?: boolean } & HTMLProps<HTMLInputElement>) {
+  const ref = React.useRef<HTMLInputElement>(null!)
+
+  React.useEffect(() => {
+    if (typeof indeterminate === 'boolean') {
+      ref.current.indeterminate = !rest.checked && indeterminate
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [ref, indeterminate])
+
+  return (
+    <input
+      type='checkbox'
+      ref={ref}
+      className={
+        className +
+        ' w-4 h-4 text-blue-600 bg-gray-100 border-gray-300 rounded focus:ring-blue-500 dark:focus:ring-blue-600 dark:ring-offset-gray-800 focus:ring-2 dark:bg-gray-700 dark:border-gray-600'
+      }
+      {...rest}
+    />
+  )
+}
+
+export { IndeterminateCheckbox }
